% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_schedule.R
\name{solar_declination}
\alias{solar_declination}
\title{Solar declination}
\usage{
solar_declination(times)
}
\arguments{
\item{times}{`POSIXct` or seconds since epoch.}
}
\value{
Declination in degrees.
}
\description{
Declination of the sun in degrees, from the standard Fourier series (Spencer 1971).
Positive in the northern summer, +/- 23.44 at the solstices, zero at the equinoxes.
}
\examples{
solar_declination(as.POSIXct(c("2021-06-21", "2021-09-22"), tz = "UTC"))
}
\seealso{
[declination_lambda_scale()]
}
