% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenithPlot.R
\name{zenithPlot}
\alias{zenithPlot}
\title{Plot simulated twilights from Lon, Lat and estimated zenith}
\usage{
zenithPlot(x, gps, threshold = 0, zenith = 95, offset = 0)
}
\arguments{
\item{x}{data frame with Light, Date columns at minimum}

\item{gps}{data frame with Date, Lon, Lat columns at minimum}

\item{threshold}{numeric Light intensity at twilight}

\item{zenith}{solar zenith angle estimated from data}

\item{offset}{offset in hours used to centre night within the plot y-axis}
}
\value{
plot of thresholded Light data with simulated zeniths highlighted
}
\description{
Plot simulated twilights from Lon, Lat and estimated zenith
}
