% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{solar_zenith}
\alias{solar_zenith}
\title{Calculate the solar zenith angle (in degrees)
Given a Unix timestamp (seconds since 1970-01-01), longitude and latitude (in degrees).}
\usage{
solar_zenith(unix_time, lon, lat)
}
\description{
Calculate the solar zenith angle (in degrees)
Given a Unix timestamp (seconds since 1970-01-01), longitude and latitude (in degrees).
}
