% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwilightFreeGrid.R
\name{TwilightFreeGrid}
\alias{TwilightFreeGrid}
\title{Reconstruct animal tracks using SGAT::essie and continuous likelihood}
\usage{
TwilightFreeGrid(
  date_time,
  light,
  grid,
  start_lat = NA_real_,
  start_lon = NA_real_,
  end_lat = NA_real_,
  end_lon = NA_real_,
  fixed = NULL,
  step_hours = 12,
  diffusion = 50,
  trans_prob = NULL,
  calibration = NULL,
  likelihood_params = NULL
)
}
\arguments{
\item{date_time}{A vector of POSIXct dates}

\item{light}{A vector of light observations}

\item{grid}{A RasterLayer or SpatialPixels object defining the grid (from TwilightFree::makeGrid)}

\item{start_lat}{Initial latitude (Decimal Degrees)}

\item{start_lon}{Initial longitude (Decimal Degrees)}

\item{end_lat}{Final latitude (optional, defaults to NA)}

\item{end_lon}{Final longitude (optional, defaults to NA)}

\item{fixed}{Optional data frame of fixed locations with columns `time` (POSIXct), `lat`, and `lon`.}

\item{step_hours}{Time step for the HMM knots in hours (default 12.0)}

\item{diffusion}{Movement diffusion in km/sqrt(day)}

\item{calibration}{Calibration parameters c(intercept, slope)}

\item{likelihood_params}{Likelihood parameters c(lambda, max_light, prob_slab)}
}
\description{
Reconstruct animal tracks using SGAT::essie and continuous likelihood
}
